+++ src/include/test/unit_api_ivf_pq_index.cc
@@ -175,7 +175,7 @@ TEST_CASE("create empty index and then train and query", "[api_ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
+  uint64_t dimensions = 3;
 
   std::string index_uri =
       (std::filesystem::temp_directory_path() / "api_ivf_pq_index").string();
@@ -247,8 +247,8 @@ TEST_CASE(
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
-  uint32_t num_subspaces = 1;
+  uint64_t dimensions = 3;
+  uint64_t num_subspaces = 1;
 
   std::string index_uri =
       (std::filesystem::temp_directory_path() / "api_ivf_pq_index").string();
@@ -498,7 +498,7 @@ TEST_CASE("storage_version", "[api_ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
+  uint64_t dimensions = 3;
 
   std::string index_uri =
       (std::filesystem::temp_directory_path() / "api_ivf_pq_index").string();
@@ -556,9 +556,9 @@ TEST_CASE("clear history with an open index", "[api_ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
+  uint64_t dimensions = 3;
   size_t n_list = 1;
-  uint32_t num_subspaces = 1;
+  uint64_t num_subspaces = 1;
   float convergence_tolerance = 0.00003f;
   size_t max_iterations = 3;
 
@@ -607,9 +607,9 @@ TEST_CASE("write and load index with timestamps", "[api_ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
+  uint64_t dimensions = 3;
   size_t n_list = 1;
-  uint32_t num_subspaces = 1;
+  uint64_t num_subspaces = 1;
   size_t max_iterations = 3;
   float convergence_tolerance = 0.00003f;
   float reassign_ratio = 0.08f;
