+++ src/include/test/unit_ivf_pq_index.cc
@@ -50,7 +50,7 @@ struct dummy_pq_index {
   auto dimensions() const {
     return 128;
   }
-  uint32_t num_subspaces() const {
+  uint64_t num_subspaces() const {
     return 16;
   }
   auto num_clusters() const {
@@ -296,7 +296,7 @@ TEST_CASE("debug w/ sk", "[ivf_pq_index]") {
 TEST_CASE("ivf_index write and read", "[ivf_pq_index]") {
   size_t dimension = 128;
   size_t nlist = 100;
-  uint32_t num_subspaces = 16;
+  uint64_t num_subspaces = 16;
   size_t max_iterationss = 4;
   size_t nprobe = 10;
   size_t k_nn = 10;
@@ -586,7 +586,7 @@ TEST_CASE("query empty index", "[ivf_pq_index]") {
   tiledb::Context ctx;
   tiledb::VFS vfs(ctx);
   size_t num_vectors = 0;
-  uint32_t dimensions = 10;
+  uint64_t dimensions = 10;
   size_t nlist = 1;
   auto index = ivf_pq_index<siftsmall_feature_type, siftsmall_ids_type>(
       nlist, dimensions / 2);
@@ -644,14 +644,14 @@ TEST_CASE("query simple", "[ivf_pq_index]") {
   tiledb::VFS vfs(ctx);
 
   size_t num_vectors = 4;
-  uint32_t dimensions = 4;
+  uint64_t dimensions = 4;
   size_t nlist = 1;
-  uint32_t num_subspaces = 2;
+  uint64_t num_subspaces = 2;
   size_t max_iterations = 1;
   float convergence_tolerance = 0.000025f;
   float reassign_ratio = 0.09f;
   std::optional<TemporalPolicy> temporal_policy = std::nullopt;
-  uint32_t num_clusters = 4;
+  uint64_t num_clusters = 4;
   using feature_type = float;
   using id_type = uint32_t;
   auto index = ivf_pq_index<feature_type, id_type>(
@@ -726,9 +726,9 @@ TEST_CASE("ivf_pq_index query index written twice", "[ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  uint32_t dimensions = 3;
+  uint64_t dimensions = 3;
   size_t n_list = 1;
-  uint32_t num_subspaces = 1;
+  uint64_t num_subspaces = 1;
   float convergence_convergence_toleranceerance = 0.00003f;
   size_t max_iterationsations = 3;
